requirements.txt
alembic==1.5.8
appdirs==1.4.4
cfgv==3.2.0
click==7.1.2
coverage==5.5
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-HTTPAuth==4.2.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
identify==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
nodeenv==1.5.0
pre-commit==2.11.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
PyYAML==5.4.1
six==1.15.0
SQLAlchemy==1.4.2
toml==0.10.2
virtualenv==20.4.3
Werkzeug==1.0.1
